fix: move render condition to MessageActions directly
